OPEN #1: Add support for gateway functionality
[Issue](http://github.com/robotmallorca/BACnet4J/issues/issue/1)

Work in progress.

- [x] Implement basic router functionality for incoming messages.
- [x] Implement basic router functionality for outgoing messages.
- [x] Implement Who-Is-Router-To-Network management.
- [x] Implement proxy-Network for non-BACNet devices.
- [-] Implement optimized transport for virtual devices.
